requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
passlib[bcrypt]==1.7.4
pyotp==2.9.0
python-jose[cryptography]==3.3.0
qrcode[pil]==7.4.2
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
requests==2.31.0
